+++ lib/assetcompute.js
@@ -15,6 +15,8 @@
 const fetch = require("@adobe/node-fetch-retry");
 const { TooManyRequestsError } = require('./error');
 const ASSET_COMPUTE_PROD_URL = "https://asset-compute.adobe.io";
+const { retry } = require("./retry");
+const clone = require('clone');
 
 /**
  * @typedef {Object} AssetComputeSource
@@ -71,7 +73,11 @@ class AssetCompute {
         this.apiKey = options.apiKey;
         this.url = options.url || ASSET_COMPUTE_PROD_URL;
         this.interval = options.interval;
+        // retry options for node-fetch-retry
         this.retryOptions = options.retryOptions || true; // default retry options
+        // custom retry options for 429 retry logic
+        this.max429RetryCount = options.max429RetryCount;
+        this.disable429Retry = options.disable429Retry;
     }
 
     /**
@@ -85,45 +91,50 @@ class AssetCompute {
      * @returns {AssetComputeRegisterResponse} Journal url
      */
     async register() {
-        const response = await fetch(`${this.url}/register`, {
-            method: "POST",
-            headers: {
-                authorization: `Bearer ${this.accessToken}`,
-                "x-gw-ims-org-id": this.org,
-                "x-ims-org-id": this.org,
-                "x-api-key": this.apiKey
-            },
-            retryOptions: this.retryOptions
-        });
-        if (!response.ok) {
-            const responseText = await response.text();
-            const message = `Unable to invoke /register: ${response.status} ${responseText}`;
-            this.handleHttpResponseNotOk(response, message);
-        } else {
-            return response.json();
-        }
+        return retry(async(self) => {
+            const response = await fetch(`${self.url}/register`, {
+                method: "POST",
+                headers: {
+                    authorization: `Bearer ${self.accessToken}`,
+                    "x-gw-ims-org-id": self.org,
+                    "x-ims-org-id": self.org,
+                    "x-api-key": self.apiKey
+                },
+                retryOptions: self.retryOptions
+            });
+            if (!response.ok) {
+                const responseText = await response.text();
+                const message = `Unable to invoke /register: ${response.status} ${responseText}`;
+                handleHttpResponseNotOk(response, message);
+            } else {
+                return response.json();
+            }
+        }, clone(this)); // to avoid `retry` changing `this`
     }
 
     /**
      * Unregister I/O events and journal
      */
     async unregister() {
-        const response = await fetch(`${this.url}/unregister`, {
-            method: "POST",
-            headers: {
-                authorization: `Bearer ${this.accessToken}`,
-                "x-gw-ims-org-id": this.org,
-                "x-ims-org-id": this.org,
-                "x-api-key": this.apiKey
+        return retry(async(self) => {
+            const response = await fetch(`${self.url}/unregister`, {
+                method: "POST",
+                headers: {
+                    authorization: `Bearer ${self.accessToken}`,
+                    "x-gw-ims-org-id": self.org,
+                    "x-ims-org-id": self.org,
+                    "x-api-key": self.apiKey
+                },
+                retryOptions: self.retryOptions
+            });
+            if (!response.ok) {
+                const responseText = await response.text();
+                const message = `Unable to invoke /unregister: ${response.status}: ${responseText} (details: ${JSON.stringify(response)})`;
+                handleHttpResponseNotOk(response, message);
+            } else {
+                return response.json();
             }
-        });
-        if (!response.ok) {
-            const responseText = await response.text();
-            const message = `Unable to invoke /unregister: ${response.status}: ${responseText} (details: ${JSON.stringify(response)})`;
-            this.handleHttpResponseNotOk(response, message);
-        } else {
-            return response.json();
-        }
+        }, clone(this)); // to avoid `retry` changing `this`
     }
 
     /**
@@ -141,49 +152,66 @@ class AssetCompute {
      * @returns {AssetComputeProcessResponse} Response with the activation id
      */
     async process(source, renditions, userData) {
-        if (Array.isArray(source)) {
-            userData = renditions;
-            renditions = source;
-            source = undefined;
-        }
-        const response = await fetch(`${this.url}/process`, {
-            method: "POST",
-            headers: {
-                authorization: `Bearer ${this.accessToken}`,
-                "x-gw-ims-org-id": this.org,
-                "x-ims-org-id": this.org,
-                "x-api-key": this.apiKey,
-                "content-type": "application/json"
-            },
-            body: JSON.stringify({
-                // TODO: providing source name, and size confuses Asset Compute so just provide the url
-                source: (source && source.url) || source,
-                renditions,
-                userData
-            }),
-            retryOptions: this.retryOptions
-        });
-        if (!response.ok) {
-            const responseText = await response.text();
-            const message = `Unable to invoke /process: ${response.status} ${responseText}`;
-            this.handleHttpResponseNotOk(response, message);
-        } else {
-            return response.json();
-        }
+        const options = {
+            source,
+            renditions,
+            userData,
+            url: this.url,
+            accessToken: this.accessToken,
+            org: this.org,
+            apiKey: this.apiKey,
+            retryOptions: this.retryOptions,
+            max429RetryCount: this.max429RetryCount,
+            disable429Retry: this.disable429Retry
+        };
+
+        return retry(async(options) => {
+            if (Array.isArray(source)) {
+                userData = renditions;
+                renditions = source;
+                source = undefined;
+            }
+            const response = await fetch(`${options.url}/process`, {
+                method: "POST",
+                headers: {
+                    authorization: `Bearer ${options.accessToken}`,
+                    "x-gw-ims-org-id": options.org,
+                    "x-ims-org-id": options.org,
+                    "x-api-key": options.apiKey,
+                    "content-type": "application/json"
+                },
+                body: JSON.stringify({
+                    // TODO: providing source name, and size confuses Asset Compute so just provide the url
+                    source: (options.source && options.source.url) || options.source,
+                    renditions: options.renditions,
+                    userData: options.userData
+                }),
+                retryOptions: options.retryOptions
+            });
+            if (!response.ok) {
+                const responseText = await response.text();
+                const message = `Unable to invoke /process: ${response.status} ${responseText}`;
+                handleHttpResponseNotOk(response, message);
+            } else {
+                return response.json();
+            }
+        }, options);
+
+        
     }
-    /**
+
+}
+/**
      * Small helper function to handle HTTP response not OK
      * @param {*} response http response
      * @param {*} errorMessage error message to throw
      */
-    handleHttpResponseNotOk(response, errorMessage) {
-        if (response.status === 429) {
-            throw new TooManyRequestsError(errorMessage, response.headers.get('retry-after'));
+function handleHttpResponseNotOk(response, errorMessage) {
+    if (response.status === 429) {
+        throw new TooManyRequestsError(errorMessage, response.headers.get('retry-after'));
 
-        }
-        throw new Error(errorMessage);
     }
-
+    throw new Error(errorMessage);
 }
 
 module.exports = {
