+++ lib/client.js
@@ -17,7 +17,6 @@ const uuid = require("uuid");
 const { AdobeAuth, AdobeIOEvents } = require("@adobe/asset-compute-events-client");
 const { AssetCompute } = require("./assetcompute");
 const { AssetComputeEventEmitter } = require("./eventemitter");
-// const { retry } = require("./client-retry");
 
 function getAssetComputeClientId(event) {
     return event.userData &&
@@ -126,7 +125,7 @@ class AssetComputeClient extends EventEmitter {
       * @param {AssetComputeIntegration} integration Asset Compute Integration
       * @param {AssetComputeClientOptions} [options=] Options provided to the client
       */
-    constructor(integration, options) {
+    constructor(integration, options={}) {
         super();
 
         // validate integration
