+++ lib/retry.js
@@ -11,7 +11,6 @@
  */
 
 "use strict";
-const { AssetComputeClient } = require("./client");
 const clone = require('clone');
 
 const MAX_RETRY_ON_HTTP_429 = 4;
@@ -19,44 +18,6 @@ const MAX_RETRY_WAIT_TIME_MS = 60000; // maximum time in ms to wait before retry
 const MIN_RETRY_WAIT_TIME_MS = 30000; // minimum time in ms to wait before retrying if no `retry-after` header is sent
 const TOO_MANY_REQUEST_ERROR = 'TooManyRequestsError';
 const TOO_MANY_REQUEST_ERROR_CODE = 429;
-/**
- * Small wrapper around AssetComputeClient adding retry logic on 429s based on `retry-after` header
- */
-class AssetComputeClientWithRetry extends AssetComputeClient {
-    constructor(integration, options) {
-        super(integration, options);
-        this.options = Object.assign(this.options || {}, {
-            max429RetryCount: (options && options.max429RetryCount || MAX_RETRY_ON_HTTP_429)
-        });
-    }
-
-    async register() {
-        return retry(async() => {
-            return super.register();
-        }, clone(this.options));
-    }
-
-    // retry on 429s
-    async unregister() {
-        return retry(async() => {
-            return super.unregister();
-        }, clone(this.options));
-    }
-
-    async process(source, renditions, userData) {
-        const options = Object.assign(
-            this.options || {},
-            {
-                source,
-                renditions,
-                userData
-            }
-        );
-        return retry(async(options) => {
-            return super.process(options.source, options.renditions, options.userData);
-        }, options);
-    }
-}
 
 /**
  * Filter out the retry options
@@ -64,11 +25,13 @@ class AssetComputeClientWithRetry extends AssetComputeClient {
  * @param {Object} options Options
  * @returns {Object} Filtered options (without retryOptions)
  */
-function filterOptions(options) {
-    if (options) {
-        delete options.max429RetryCount;
+function filterOptions(options={}) {
+    const newOpts = clone(options);
+    if (newOpts) {
+        delete newOpts.max429RetryCount;
+        delete newOpts.disable429Retry;
     }
-    return options;
+    return newOpts;
 }
 
 
@@ -98,6 +61,10 @@ function retryWaitTime(retryAfter) {
      * @returns 
      */
 function shouldRetry(attempt, error, retryOpts) {
+    // if retry is disabled, return false
+    if (retryOpts.disable429Retry) {
+        return false;
+    }
     if (error.code === TOO_MANY_REQUEST_ERROR_CODE && error.name === TOO_MANY_REQUEST_ERROR) {
         return attempt < retryOpts.max429RetryCount;
     } else {
@@ -146,13 +113,13 @@ async function retryInvoke(asyncFunc, options, retryOpts) {
  */
 async function retry(asyncFunc, options) {
     const retryOpts = {
-        max429RetryCount: options.max429RetryCount || MAX_RETRY_ON_HTTP_429
+        max429RetryCount: options && options.max429RetryCount || MAX_RETRY_ON_HTTP_429,
+        disable429Retry: options && options.disable429Retry || false
     };
     options = filterOptions(options); // remove retry options from options passed to actual fetch
     return retryInvoke(asyncFunc, options, retryOpts);
 }
 
 module.exports = {
-    AssetComputeClientWithRetry,
     retry
 };
